requirements.txt
requests==2.25.1
PyYAML==6.0.1
lxml==4.9.3
